+++ src/components/pages/article-page/article-page.js
@@ -1,4 +1,4 @@
-import { Button, ButtonGroup, Chip, CircularProgress, Container, Link, Typography } from "@material-ui/core";
+import { Button, Chip, CircularProgress, Container, Link, Typography } from "@material-ui/core";
 import React, { useState, useEffect } from "react";
 import { makeStyles } from '@material-ui/core/styles';
 import { CardHeader, Avatar } from "@material-ui/core";
@@ -7,12 +7,12 @@ import AddIcon from '@material-ui/icons/Add';
 import EditIcon from '@material-ui/icons/Edit';
 import DeleteIcon from '@material-ui/icons/Delete';
 import { Link as RouterLink, Redirect } from 'react-router-dom';
+import { connect } from "react-redux";
 
-import { API, fetchArticles } from "../../../services/requst";
+import { API } from "../../../services/requst";
 import { FormatData } from "../../format-date";
 import { withRouter } from "react-router-dom";
 import AddComments from "../../add-comments/add-comments";
-import { connect } from "react-redux";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -77,7 +77,7 @@ const useStyles = makeStyles((theme) => ({
     styleTag: {
         marginRight: theme.spacing(1),
     },
-    linkRemove: {
+    linkEdit: {
         color: "grey",
         "&:hover": {
             color: "white",
@@ -120,12 +120,26 @@ const useStyles = makeStyles((theme) => ({
             transition: "0.5s",
             border: "1px solid green",
         },
+    },
+    rootLinkLogReg: {
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+    },
+    linkLogin: {
+        color: "green",
+        fontWeight: "bold",
+        marginRight: theme.spacing(1),
+    },
+    linkRegister: {
+        color: "green",
+        fontWeight: "bold",
+        marginLeft: theme.spacing(1),
+        marginRight: theme.spacing(1),
     }
 }));
 
-const token = localStorage.getItem("diplomaToken");
-
-const ArticlePage = ({ match, usernameAutorization }) => {
+const ArticlePage = ({ match, usernameAutorization, token }) => {
     const { id } = match.params;
 
     const classes = useStyles();
@@ -136,6 +150,7 @@ const ArticlePage = ({ match, usernameAutorization }) => {
     const [removeArticle, setRemoveArticle] = useState(false);
     const [favoritedArticle, setFavoritedArticle] = useState(false);
     const [countFavoritedArticle, setCountFavoritedArticle] = useState("");
+    const [redirectPage, setRedirectPage] = useState(false);
 
     useEffect(() => {
         async function articleFetchData() {
@@ -150,7 +165,7 @@ const ArticlePage = ({ match, usernameAutorization }) => {
             }
         }
         articleFetchData();
-    }, [setFollowingAuthor, setFavoritedArticle, setCountFavoritedArticle]);
+    }, [setFollowingAuthor, setFavoritedArticle, setCountFavoritedArticle, id]);
 
     const hendelRemoveArticle = async () => {
         await API.delete(`https://conduit.productionready.io/api/articles/${id}`);
@@ -158,34 +173,44 @@ const ArticlePage = ({ match, usernameAutorization }) => {
     }
 
     const hendelFollow_UnFollow = async () => {
-        if (followingAuthor) {
-            const unFollow = await API.delete(`https://conduit.productionready.io/api/profiles/${authorArticle}/follow`);
-            setFollowingAuthor(unFollow.data.profile.following);
+        if (token) {
+            if (followingAuthor) {
+                const unFollow = await API.delete(`https://conduit.productionready.io/api/profiles/${authorArticle}/follow`);
+                setFollowingAuthor(unFollow.data.profile.following);
+            }
+            else {
+                const follow = await API.post(`https://conduit.productionready.io/api/profiles/${authorArticle}/follow`);
+                setFollowingAuthor(follow.data.profile.following);
+            }
         }
         else {
-            const follow = await API.post(`https://conduit.productionready.io/api/profiles/${authorArticle}/follow`);
-            setFollowingAuthor(follow.data.profile.following);
+            setRedirectPage(true);
         }
     }
 
     const hendelFavorite = async () => {
-        if (favoritedArticle) {
-            const unFavorite = await API.delete(`https://conduit.productionready.io/api/articles/${id}/favorite`);
-            setFavoritedArticle(unFavorite.data.article.favorited);
-            setCountFavoritedArticle(unFavorite.data.article.favoritesCount);
+        if (token) {
+            if (favoritedArticle) {
+                const unFavorite = await API.delete(`https://conduit.productionready.io/api/articles/${id}/favorite`);
+                setFavoritedArticle(unFavorite.data.article.favorited);
+                setCountFavoritedArticle(unFavorite.data.article.favoritesCount);
+            }
+            else {
+                const favorite = await API.post(`https://conduit.productionready.io/api/articles/${id}/favorite`);
+                setFavoritedArticle(favorite.data.article.favorited);
+                setCountFavoritedArticle(favorite.data.article.favoritesCount);
+            }
         }
         else {
-            const favorite = await API.post(`https://conduit.productionready.io/api/articles/${id}/favorite`);
-            setFavoritedArticle(favorite.data.article.favorited);
-            setCountFavoritedArticle(favorite.data.article.favoritesCount);
+            setRedirectPage(true);
         }
     }
 
     const yourArticleRender = () => {
         return (
             <div className={classes.btnRoot}>
-                <Link component={RouterLink} to={`/edit_article/${id}`} startIcon={<EditIcon />} className={classes.linkRemove}>
-                    <Button variant="outlined" className={classes.btnOne}>
+                <Link component={RouterLink} to={`/edit_article/${id}`} className={classes.linkEdit}>
+                    <Button variant="outlined" className={classes.btnOne} startIcon={<EditIcon />}>
                         Edit Article
                     </Button>
                 </Link>
@@ -200,22 +225,28 @@ const ArticlePage = ({ match, usernameAutorization }) => {
         //console.log("####: countFavoritedArticle", countFavoritedArticle);
         return (
             <div className={classes.btnRoot}>
-                <Button
-                    variant="outlined"
-                    startIcon={<AddIcon />}
-                    className={followingAuthor ? classes.followStyleBtnActive : classes.followStyleBtn}
-                    onClick={hendelFollow_UnFollow}
-                >
-                    {followingAuthor ? `UnFollow ${authorArticle}` : `Follow ${authorArticle}`}
-                </Button>
-                <Button
-                    variant="outlined"
-                    className={favoritedArticle ? classes.favoriteStyleBtnActive : classes.favoriteStyleBtn}
-                    startIcon={<FavoriteIcon />}
-                    onClick={hendelFavorite}
-                >
-                    Favorite Article ({countFavoritedArticle})
-                </Button>
+                {redirectPage
+                    ? <Redirect to="/login" />
+                    :
+                    <>
+                        <Button
+                            variant="outlined"
+                            startIcon={<AddIcon />}
+                            className={followingAuthor ? classes.followStyleBtnActive : classes.followStyleBtn}
+                            onClick={hendelFollow_UnFollow}
+                        >
+                            {followingAuthor ? `UnFollow ${authorArticle}` : `Follow ${authorArticle}`}
+                        </Button>
+                        <Button
+                            variant="outlined"
+                            className={favoritedArticle ? classes.favoriteStyleBtnActive : classes.favoriteStyleBtn}
+                            startIcon={<FavoriteIcon />}
+                            onClick={hendelFavorite}
+                        >
+                            Favorite Article ({countFavoritedArticle})
+                        </Button>
+                    </>
+                }
             </div>
         )
     }
@@ -255,9 +286,30 @@ const ArticlePage = ({ match, usernameAutorization }) => {
                             ? data.tagList.map((tag, index) => <Chip label={tag} key={index} className={classes.styleTag} />)
                             : null}
                         <hr />
-                        {token ? <AddComments slug={id} /> : null}
+                        {token
+                            ? <AddComments slug={id} />
+                            :
+                            <div className={classes.rootLinkLogReg}>
+                                <Link
+                                    component={RouterLink}
+                                    to={`/login`}
+                                    className={classes.linkLogin}
+                                >
+                                    Sign in
+                                </Link>
+                                    or
+                                <Link
+                                    component={RouterLink}
+                                    to={`/register`}
+                                    className={classes.linkRegister}
+                                >
+                                    sign up
+                                </Link>
+                                    to add comments on this article.
+                            </div>
+                        }
                     </Container>
-                    
+
                     {removeArticle ? <Redirect to="/" /> : null}
                 </>
             }
@@ -266,8 +318,10 @@ const ArticlePage = ({ match, usernameAutorization }) => {
 }
 
 const mapStateToProps = (state) => {
+    const { token, username } = state.user
     return {
-        usernameAutorization: state.user.username,
+        token,
+        usernameAutorization: username,
     }
 }
